collector: now with protean ProcessStream().  appears to properly advance targets and persist at appropriate intervals.  Now to fix promoteTarget().

+++ collector/tools.go
@@ -110,7 +110,7 @@ func cleanFile(fileName string, contents []byte, _type string) {
 		if err != nil {
 			panic(err)
 		}
-		near, _ := isNear(int64(time1), time2, int64(45))
+		near, _ := isNear(int64(time1), time2, 45)
 		if near {
 			good += 1
 			funcs.LazyAppendFile(filepath.Dir(cleanFilename), filepath.Base(cleanFilename), string(row))
